Restructure ConsoleDispatcher file location
This commit moves the ConsoleDispatcher underneath HttpDispatchers so
that it is clearer that there should be room for growth for other kinds
of http traffic objects for other services, such as the Puppet Server
and PuppetDB. There are also various changes to Rdoc strings.
