+++ lib/scooter/httpdispatchers.rb
@@ -1,204 +1,12 @@
-%w( classifier rbac ).each do |lib|
+%w( consoledispatcher ).each do |lib|
   require "scooter/httpdispatchers/#{lib}"
 end
 
 module Scooter
-
+  # This module is just the housing for the single dispatcher we have right now,
+  # ConsoleDispatcher, but should eventually include other dispatchers for other
+  # services that talk over http, such as the Puppet Server and PuppetDB.
   module HttpDispatchers
-
-    # == Quick-start guide
-    # === Quick example creating an api dispatcher and ui dispatcher
-    #
-    #  require 'scooter/httpdispatchers'
-    #
-    #  api = ConsoleDispatcher.new(dashboard)
-    #
-    #  #creates a new user and returns a new ConsoleDispatcher
-    #  new_user = api.create_local_user('login' => 'new_guy')
-    #  new_user.signin
-    #
-    #  #this will trigger the middleware error handler, because you cannot
-    #  #create two users with the same login
-    #  api.create_local_user('login' => 'new_guy')
-    #
-    #  #this will return the new_guy's data
-    #  api.get_console_dispatcher_data(new_guy)
-    #
-    #  api.delete_local_console_dispatcher(new_user)
-    #
-    #  #the same method will now return nil because new_user is deleted
-    #  api.get_console_dispatcher_data(new_guy)
-    class ConsoleDispatcher
-
-      include Scooter::HttpDispatchers::Rbac
-      include Scooter::HttpDispatchers::Classifier
-      attr_accessor :connection, :dashboard, :credentials, :ssl
-      Credentials = Struct.new(:login, :password)
-
-      # This class is designed to interact with any of the pe-console-services:
-      # RBAC, Node Classifier, and the Activity Service. The most common use
-      # case will be for tests executed with beaker, for which you will want
-      # to pass in the dashboard for initialization. The other parameter that
-      # you can pass in optionally is credentials. Passing in credentials is
-      # essential for building the correct routes if you are using RBAC users,
-      # local users, or the default admin account. If no credentials are passed
-      # in, the dispatcher defaults to using certificates to connect to the
-      # API's directly.
-      #
-      # If you are using this class outside of beaker, you can initialize an
-      # object without any parameters and supply your own dashboard as a String,
-      # credentials, Faraday connection object, and ssl parameters. Note that
-      # this is largely untested. Use outside of a Beaker test run at your own
-      # risk.
-      # @param dashboard [Unix::Host optional] the dashboard specified in your
-      #   beaker config. If no dashboard is passed, then the initialization
-      #   process is skipped and you must define the connection and dashboard
-      #   manually.
-      #
-      # @param credentials(Hash optional) Provide credentials if you wish to
-      #   communicate through the UI proxy. If no credentials are provided, then
-      #   it is assumed dispatching object will send traffic directly to the
-      #   APIs.
-      def initialize(dashboard = nil, credentials=nil)
-        @dashboard = dashboard if dashboard.is_a?(Unix::Host)
-        @credentials = Credentials.new(credentials[:login],
-                                       credentials[:password]) if credentials
-
-        @connection = create_default_connection_and_initialize if @dashboard
-      end
-
-      def create_default_connection
-        Faraday.new do |conn|
-          conn.request :json
-
-          # This logger will need to be configurable somehow..., maybe based on
-          # beaker log-level?
-          conn.response :logger, nil, bodies: true
-          conn.response :follow_redirects
-          conn.response :json, :content_type => /\bjson$/
-          conn.response :raise_error
-
-          conn.use :cookie_jar
-
-          conn.adapter :net_http
-        end
-      end
-
-      def create_default_connection_and_initialize
-        connection = create_default_connection
-        set_host_and_port(connection)
-        acquire_ssl_components
-        add_ssl_components_to_connection(connection)
-        connection
-      end
-
-
-      # If you would like to run tests that expect 400 or even 500 responses,
-      # apply this method to remove the <tt>:raise_error</tt> middleware.
-      def remove_error_checking(connection=@connection)
-        connection.builder.delete(Faraday::Response::RaiseError)
-      end
-
-      def set_host_and_port(connection=@connection)
-        connection.url_prefix.scheme = 'https'
-        connection.url_prefix.host = "#{@dashboard}"
-
-        if is_certificate_dispatcher?
-          connection.url_prefix.port = 4433
-        else
-          connection.url_prefix.port = 443
-        end
-      end
-
-      def set_classifier_path(connection=@connection)
-        if is_certificate_dispatcher?
-          connection.url_prefix.path = '/classifier-api'
-        else
-          connection.url_prefix.path = '/node_groups/classifier-api'
-        end
-      end
-
-      def set_rbac_path(connection=@connection)
-        if is_certificate_dispatcher?
-          connection.url_prefix.path = '/rbac-api'
-        else
-          connection.url_prefix.path = '/rbac/rbac-api'
-        end
-      end
-
-      def set_activity_service_path(connection=@connection)
-        if is_certificate_dispatcher?
-          connection.url_prefix.path = '/activity-api'
-        else
-          connection.url_prefix.path = '/rbac/activity-api'
-        end
-      end
-
-      def acquire_ssl_components
-        if !@dashboard.is_a?(Unix::Host)
-          raise 'Can only acquire SSL certs if the dashboard is a Unix::Host'
-        end
-        @ssl = {}
-        @ssl['ca_file'] = Scooter::Utilities::BeakerUtilities.pe_ca_cert_file(@dashboard)
-        if is_certificate_dispatcher?
-          client_key = Scooter::Utilities::BeakerUtilities.pe_private_key(@dashboard)
-          client_cert = Scooter::Utilities::BeakerUtilities.pe_hostcert(@dashboard)
-
-          @ssl['client_key']  = OpenSSL::PKey.read(client_key)
-          @ssl['client_cert'] = OpenSSL::X509::Certificate.new(client_cert)
-        end
-      end
-
-      def add_ssl_components_to_connection(connection=@connection)
-        @ssl.each do |k, v|
-          connection.ssl[k] = v
-        end
-      end
-
-      # Run this method if you have replaced the connection with a different
-      # Faraday connection; it will just set the prefix_url and add the ssl
-      # components already stored. This method is used primarily when you have
-      # not passed in a dashboard from beaker and are acquiring ssl credentials
-      # through separate means; it skips the <tt>acquire_ssl_components</tt>
-      # and assumes that you acquired an ssl client_key and client_cert through
-      # separate means.
-      def reinitialize_connection(connection=@connection)
-        set_host_and_port(connection)
-        add_ssl_components_to_connection(connection)
-      end
-
-      def is_certificate_dispatcher?
-        true unless @credentials
-      end
-
-      def signin(login=@credentials.login, password=@credentials.password)
-        response = @connection.post "/auth/login" do |request|
-          request.headers['Content-Type'] = 'application/x-www-form-urlencoded; charset=UTF-8'
-          request.body = "username=#{login}&password=#{CGI.escape(password)}"
-        end
-
-        #return the response if the status code was not 200
-        return response if response.status != 200
-
-        # Acquire the X-CSRF token if signin was successful. The console doesn't
-        # require the token for every request, but it is required for just
-        # about every non-GET request. This could be smarter to only send on
-        # the required requests, but doesn't seem necessary to embed that logic;
-        # we will just send it on every request after successful signin.
-        response_body = @connection.get('/rbac/#/users').env.body
-        parsed_body = Nokogiri::HTML(response_body)
-        token = parsed_body.css("meta[name='__anti-forgery-token']")[0].attributes['content'].value
-        @connection.headers['X-CSRF-Token'] = token
-      end
-
-      def reset_local_user_password(token, new_password)
-        @connection.post "https://#{dashboard}/auth/reset" do |request|
-          request.headers['Content-Type'] = 'application/x-www-form-urlencoded; charset=UTF-8'
-          request.body = "password=#{new_password}&token=#{token}"
-        end
-
-      end
-    end
   end
 end
